Move survey submission to more logical location
